requirements.txt
Flask>=2.0.0
transformers>=4.25.0
torch>=1.13.0
torchvision>=0.14.0
Pillow>=9.0.0
Werkzeug>=2.0.0